+++ packages/react-widgets/src/widgets/HistogramWidget.js
@@ -22,8 +22,9 @@ const EMPTY_ARRAY = [];
  * @param  {string} props.title - Title to show in the widget header.
  * @param  {string} props.dataSource - ID of the data source to get the data from.
  * @param  {string} props.column - Name of the data source's column to get the data from.
- * @param  {string} props.operation - Operation to apply to the column. Must be one of those defined in `AggregationTypes` object.
- * @param  {number[]} props.ticks - Array of thresholds for the X axis.
+ * @param  {string} [props.operation] - Operation to apply to the column. Must be one of those defined in `AggregationTypes` object.
+ * @param  {number[]} [props.ticks] - Array of thresholds for the X axis.
+ * @param  {number} [props.bins] - Number of bins to calculate the ticks.
  * @param  {Function} [props.xAxisformatter] - Function to format X axis values.
  * @param  {Function} [props.formatter] - Function to format Y axis values.
  * @param  {boolean} [props.tooltip=true] - Whether to show a tooltip or not
@@ -35,42 +36,50 @@ const EMPTY_ARRAY = [];
  * @param  {Object} [props.noDataAlertProps] - Extra props to pass to [NoDataAlert]()
  * @param  {Object} [props.droppingFeaturesAlertProps] - Extra props to pass to [NoDataAlert]() when dropping feature
  */
-function HistogramWidget(props) {
-  const {
-    id,
-    title,
-    dataSource,
-    column,
-    operation,
-    ticks,
-    xAxisFormatter,
-    dataAxis,
-    formatter,
-    tooltip,
-    animation,
-    filterable,
-    global,
-    onError,
-    wrapperProps,
-    noDataAlertProps,
-    droppingFeaturesAlertProps = defaultDroppingFeaturesAlertProps
-  } = props;
+function HistogramWidget({
+  id,
+  title,
+  dataSource,
+  column,
+  operation,
+  ticks: _ticks,
+  xAxisFormatter,
+  bins,
+  formatter,
+  tooltip,
+  tooltipFormatter,
+  animation,
+  filterable,
+  global,
+  onError,
+  wrapperProps,
+  noDataAlertProps,
+  droppingFeaturesAlertProps = defaultDroppingFeaturesAlertProps
+}) {
   const dispatch = useDispatch();
   const isDroppingFeatures = useSelector((state) =>
     checkIfSourceIsDroppingFeature(state, dataSource)
   );
 
-  const { data = [], isLoading } = useWidgetFetch(getHistogram, {
+  const { data: _data, isLoading } = useWidgetFetch(getHistogram, {
     id,
     dataSource,
     params: {
       column,
       operation,
-      ticks
+      ticks: _ticks,
+      bins
     },
     global,
     onError
   });
+
+  const {
+    min = Number.MIN_SAFE_INTEGER,
+    max = Number.MAX_SAFE_INTEGER,
+    data = [],
+    ticks = _ticks
+  } = _data || {};
 
   const thresholdsFromFilters = useWidgetFilterValues({
     dataSource,
@@ -94,23 +103,10 @@ function HistogramWidget(props) {
       .filter((v) => v !== null);
   }, [thresholdsFromFilters, ticks]);
 
-  const tooltipFormatter = useCallback(
-    ([serie]) => {
-      const formattedValue = formatter
-        ? formatter(serie.value, serie.dataIndex, ticks)
-        : { prefix: '', value: serie.value };
-
-      return typeof formattedValue === 'object'
-        ? `${formattedValue.prefix}${formattedValue.value}`
-        : formattedValue;
-    },
-    [formatter, ticks]
-  );
-
   const handleSelectedBarsChange = useCallback(
-    ({ bars }) => {
-      if (bars && bars.length) {
-        const thresholds = bars.map((i) => {
+    (selectedBars) => {
+      if (selectedBars?.length) {
+        const thresholds = selectedBars.map((i) => {
           let left = ticks[i - 1];
           let right = ticks.length !== i ? ticks[i] : undefined;
 
@@ -137,22 +133,14 @@ function HistogramWidget(props) {
     [column, dataSource, id, dispatch, ticks]
   );
 
-  const ticksForDataAxis = ticks.reduce((acc, tick, i) => {
-    if (acc.length === 0) {
-      return [`< ${tick}`];
-    }
-    if (i === ticks.length - 1) {
-      return [...acc, `< ${tick}`, `>= ${tick}`];
-    }
-    return [...acc, `< ${tick}`];
-  }, []);
-
   return (
     <WrapperWidgetUI title={title} {...wrapperProps} isLoading={isLoading}>
       {(data.length && !isDroppingFeatures) || isLoading ? (
         <HistogramWidgetUI
           data={data}
-          dataAxis={dataAxis || ticksForDataAxis}
+          min={min}
+          max={max}
+          ticks={ticks}
           selectedBars={selectedBars}
           onSelectedBarsChange={handleSelectedBarsChange}
           tooltip={tooltip}
@@ -176,14 +164,15 @@ HistogramWidget.propTypes = {
   title: PropTypes.string.isRequired,
   dataSource: PropTypes.string.isRequired,
   column: PropTypes.string.isRequired,
-  operation: PropTypes.oneOf(Object.values(AggregationTypes)).isRequired,
+  ticks: PropTypes.arrayOf(PropTypes.number),
+  bins: PropTypes.number,
+  operation: PropTypes.oneOf(Object.values(AggregationTypes)),
   xAxisFormatter: PropTypes.func,
   formatter: PropTypes.func,
   tooltip: PropTypes.bool,
   animation: PropTypes.bool,
   filterable: PropTypes.bool,
   global: PropTypes.bool,
-  ticks: PropTypes.array.isRequired,
   onError: PropTypes.func,
   wrapperProps: PropTypes.object,
   noDataAlertProps: PropTypes.object,
@@ -191,6 +180,9 @@ HistogramWidget.propTypes = {
 };
 
 HistogramWidget.defaultProps = {
+  bins: 15,
+  ticks: [],
+  operation: AggregationTypes.COUNT,
   tooltip: true,
   animation: true,
   filterable: true,